feat: 完善系统管理模块，修复分页total序列化bug，合并SQL文件，清理临时文件

+++ vben-admin-backend/java-backend/src/main/java/com/vben/backend/common/result/PageResult.java
@@ -16,10 +16,10 @@ public class PageResult<T> {
     private List<T> items;
 
     /** 总条数 */
-    private long total;
+    private int total;
 
     public PageResult(List<T> items, long total) {
         this.items = items;
-        this.total = total;
+        this.total = (int) total;
     }
 }
